# dkquantweb/dkquantweb/OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "描述：用于管理集团旗下公司的人员信息,具体包括XXX,XXX模块...",
    "version": "版本号:3.8.9",
    "title": "标题：若依管理系统_接口文档",
    "contact": {
      "name": "RuoYi"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/dev-api/strategy/fun_fund/{id}": {
      "get": {
        "tags": [
          "资金方法相关接口"
        ],
        "summary": "获取资金方法详细信息",
        "operationId": "getInfoUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/fun_position/{id}": {
      "get": {
        "tags": [
          "开仓方法相关接口"
        ],
        "summary": "获取开仓方法详细信息",
        "operationId": "getInfoUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/fun_profitLoss/{id}": {
      "get": {
        "tags": [
          "止盈止损方法相关接口"
        ],
        "summary": "获取止盈止损方法详细信息",
        "operationId": "getInfoUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/fun_selection/{id}": {
      "get": {
        "tags": [
          "选股方法相关接口"
        ],
        "summary": "获取选股方法详细信息",
        "operationId": "getInfoUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_fund/{id}": {
      "get": {
        "tags": [
          "资金策略相关接口"
        ],
        "summary": "获取资金策略详细信息",
        "operationId": "getInfoUsingGET_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_position/{id}": {
      "get": {
        "tags": [
          "开仓策略相关接口"
        ],
        "summary": "获取开仓策略详细信息",
        "operationId": "getInfoUsingGET_7",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_profitLoss/{id}": {
      "get": {
        "tags": [
          "止盈止损策略相关接口"
        ],
        "summary": "获取止盈止损策略详细信息",
        "operationId": "getInfoUsingGET_8",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_pv/{id}": {
      "get": {
        "tags": [
          "私有策略相关接口"
        ],
        "summary": "获取私有策略详细信息",
        "operationId": "getInfoUsingGET_9",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/strategy/strategy_selection/{id}": {
      "get": {
        "tags": [
          "选股策略相关接口"
        ],
        "summary": "获取选股策略详细信息",
        "operationId": "getInfoUsingGET_10",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/account/{id}": {
      "get": {
        "tags": [
          "策略账户信息相关接口"
        ],
        "summary": "获取策略账户信息详细信息",
        "operationId": "getInfoUsingGET_11",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/args/{id}": {
      "get": {
        "tags": [
          "策略参数信息相关接口"
        ],
        "summary": "获取策略参数信息详细信息",
        "operationId": "getInfoUsingGET_12",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/orders/{id}": {
      "get": {
        "tags": [
          "策略订单信息相关接口"
        ],
        "summary": "获取策略订单信息详细信息",
        "operationId": "getInfoUsingGET_13",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/position/{id}": {
      "get": {
        "tags": [
          "策略仓位相关接口"
        ],
        "summary": "获取策略仓位详细信息",
        "operationId": "getInfoUsingGET_14",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/report/{id}": {
      "get": {
        "tags": [
          "策略绩效指标相关接口"
        ],
        "summary": "获取策略绩效指标详细信息",
        "operationId": "getInfoUsingGET_6",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/strategy/export": {
      "post": {
        "tags": [
          "策略相关接口"
        ],
        "summary": "导出策略列表",
        "operationId": "exportUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "argsId",
            "in": "query",
            "description": "参数",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "createBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "delFlag",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "策略id",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxPosition",
            "in": "query",
            "description": "最多持仓",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "owner",
            "in": "query",
            "description": "归属(用户id)",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "params",
            "in": "query",
            "required": false,
            "type": "object"
          },
          {
            "name": "remark",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "searchValue",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "strategyName",
            "in": "query",
            "description": "策略名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "updateBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "updateTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/strategy/list": {
      "get": {
        "tags": [
          "策略相关接口"
        ],
        "summary": "查询策略列表",
        "operationId": "listUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "argsId",
            "in": "query",
            "description": "参数",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "createBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "delFlag",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "策略id",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxPosition",
            "in": "query",
            "description": "最多持仓",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "owner",
            "in": "query",
            "description": "归属(用户id)",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "params",
            "in": "query",
            "required": false,
            "type": "object"
          },
          {
            "name": "remark",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "searchValue",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "strategyName",
            "in": "query",
            "description": "策略名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "updateBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "updateTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TableDataInfo",
              "originalRef": "TableDataInfo"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/dev-api/system/strategy/{ids}": {
      "delete": {
        "tags": [
          "策略相关接口"
        ],
        "summary": "删除策略",
        "operationId": "removeUsingDELETE_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "description": "ids",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    },
    "/dev-api/system/strategy/{id}": {
      "get": {
        "tags": [
          "策略相关接口"
        ],
        "summary": "获取策略详细信息",
        "operationId": "getInfoUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Authorization": [
              "global"
            ]
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "TableDataInfo": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "rows": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "TableDataInfo"
    },
...
  }
}